integrate first geometry-tessellation test (with browserify fixes)
