+++ package.json
@@ -41,7 +41,6 @@
     "gulp-replace": "^0.4.0",
     "mocha": "^1.21.5",
     "rfolderify": "^1.3.0",
-    "unpathify": "^0.1.4",
     "vinyl-transform": "0.0.1"
   }
 }
