+++ test/browser/fake-chai.js
@@ -1,0 +1,4 @@
+/* jshint node: true */
+
+// stub to let the browserified tests use the page-provided Chai
+module.exports = window.chai;
