+++ test/index.html
@@ -14,7 +14,8 @@
       mocha.setup('tdd');
     </script>
 
-    <script src="tests_browserified.js"></script>
+    <script src="../libtess.min.js"></script>
+    <script src="tests-browserified.js"></script>
 
     <script>
       mocha.run();
